requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
Django==3.2
django-debug-toolbar==3.2.4
django-elasticsearch-dsl==7.2.1
django-elasticsearch-dsl-drf==0.18
django-nine==0.2.5
djangorestframework==3.13.1
elasticsearch==7.16.1
elasticsearch-dsl==7.4.0
psycopg2==2.9.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
